feat(build): add workflow for building .NET projects
This commit adds a new workflow file `build-dotnet.yml` that sets up a build process for .NET projects. The workflow includes steps to checkout code, setup the specified version of .NET, install NBGV (if enabled), restore dependencies, build the project(s), test the project(s) and save test results as artifacts. The build artifact is also uploaded at the end of the workflow.